requirements.txt
flask
python-dotenv
PyPDF2
python-docx
langchain
langchain-community
langchain-openai
langchain-huggingface
sentence-transformers
requests
faiss-cpu
python-dateutil
fitz
pytesseract
PIL